Refactor indeed to use Scraper_Engine
